Move `registry_models` to `schema`
